chore: move import and export to extension file
