+++ packages/cli/src/commands/snippets.ts
@@ -114,71 +114,11 @@ module.exports = {
       authToken
     );
 
-    // Import all snippets through the API and recursively merge them with our local snippets
-    const importSnippets = async () => {
-      toolbox.print.info("Importing Snippets from Shopware");
-
-      const { keepLocal } = inputParameters;
-
-      if (!keepLocal) {
-        toolbox.print.warning(
-          "Local Snippets will be overridden (run with --keep-local to keep local changes)"
-        );
-      }
-
-      // For every locale
-      for (let [locale, snippetSetIdentifier] of Object.entries(
-        snippetSetsMap
-      )) {
-        // Fetch snippets from API
-        let flatRemoteSnippets = [];
-        try {
-          flatRemoteSnippets = await toolbox.snippets.fetchFromApi(
-            inputParameters.shopwareEndpoint,
-            authToken,
-            snippetSetIdentifier
-          );
-        } catch (error) {
-          toolbox.print.error(`Fetch from API error: ${error}`);
-        }
-
-        // Make them objects that we can write to locale files
-        const remoteSnippets = toolbox.snippets.inflateSnippetObject(
-          flatRemoteSnippets,
-          {}
-        );
-
-        // Generate locale file path
-        const localSnippetsPath = path.join("locales", `${locale}.json`);
-
-        if (!toolbox.filesystem.exists(localSnippetsPath)) {
-          toolbox.print.warning(
-            `Creating '${localSnippetsPath}, have you deleted your local snippets?`
-          );
-          toolbox.filesystem.write(localSnippetsPath, {});
-        }
-        const localSnippets = require(localSnippetsPath);
-
-        var mergedSnippets = {};
-        const { merge } = require("lodash");
-
-        if (keepLocal) {
-          // First apply remote ones and then apply local ones
-          mergedSnippets = merge(remoteSnippets.pwa, localSnippets);
-        } else {
-          mergedSnippets = merge(localSnippets, remoteSnippets.pwa);
-        }
-
-        // Merge old and new and write them into the local snippets file
-        await toolbox.filesystem.writeAsync(localSnippetsPath, mergedSnippets);
-
-        toolbox.print.success(
-          `Wrote ${flatRemoteSnippets.length} snippets to '${localSnippetsPath}'.`
-        );
-      }
-    };
-
-    await importSnippets();
+    await toolbox.snippets.importSnippets({
+      inputParameters,
+      snippetSetsMap,
+      authToken,
+    });
 
     const doExport = inputParameters.export;
 
@@ -189,38 +129,10 @@ module.exports = {
       return;
     }
 
-    // Export snippets from locales files and send them to the API
-    const exportSnippets = async () => {
-      toolbox.print.info("Exporting Snippets to Shopware");
-
-      const localesPath = "locales";
-      const files = await toolbox.filesystem.listAsync(localesPath);
-
-      for (let i = 0; i < files.length; i++) {
-        const snippetSet = require(path.join(localesPath, files[i]));
-
-        const flatSnippets = toolbox.snippets.flattenSnippetObject(
-          snippetSet,
-          "pwa"
-        );
-
-        // Get locale from file name
-        let locale = files[i].split(".")[0];
-
-        // Get correct snippet set id for that locale
-        try {
-          toolbox.snippets.writeToApi(
-            flatSnippets,
-            snippetSetsMap[locale],
-            inputParameters.shopwareEndpoint,
-            authToken
-          );
-        } catch (error) {
-          toolbox.print.error(`Export to API error: ${error}`);
-        }
-      }
-    };
-
-    await exportSnippets();
+    await toolbox.snippets.exportSnippets({
+      inputParameters,
+      snippetSetsMap,
+      authToken,
+    });
   },
 };
